:truck: Functions organization
